requirements.txt
cssselect==0.9.2
Flask==0.10.1
lxml==3.6.0
PyYAML==3.11
query-string==0.0.5
requests==1.2.3
twilio==5.4.0
